281: create look/around.js
